requirements.txt
flask
flask_cors
python-dotenv
sqlalchemy
PyJWT
langchain
langchain_core
langchain_community
langchain_groq
google-genai
scikit-learn
duckduckgo_search
psycopg2-binary
gunicorn
faiss-cpu
sentence-transformers
tenacity
requests
beautifulsoup4
unstructured
python-docx
pypdf